+++ lua/ergoterm/config.lua
@@ -21,6 +21,7 @@ M.NULL_CALLBACK = function(...) end
 ---@alias default_action fun(term: Terminal)
 ---@alias on_close fun(term: Terminal)
 ---@alias on_focus fun(term: Terminal)
+---@alias on_unfocus fun(term: Terminal)
 ---@alias on_job_exit fun(t: Terminal, job: number, exit_code: number, event: string)
 ---@alias on_job_stdout fun(t: Terminal, channel_id: number, data: string[], name: string)
 ---@alias on_job_stderr fun(t: Terminal, channel_id: number, data: string[], name: string)
@@ -58,6 +59,7 @@ M.NULL_CALLBACK = function(...) end
 ---@field float_winblend number?
 ---@field on_close on_close?
 ---@field on_focus on_focus?
+---@field on_unfocus on_unfocus?
 ---@field on_job_exit on_job_exit?
 ---@field on_job_stdout on_job_stdout?
 ---@field on_job_stderr on_job_stderr?
@@ -121,6 +123,7 @@ local config = {
     },
     on_close = M.NULL_CALLBACK,
     on_focus = M.NULL_CALLBACK,
+    on_unfocus = M.NULL_CALLBACK,
     on_job_exit = M.NULL_CALLBACK,
     on_open = M.NULL_CALLBACK,
     on_stop = M.NULL_CALLBACK,